Move controllers to api folder
